Test that the stream management negotiation works.
This adds implements the basic server side component
and adds a way to indicate that the test should not have
succeeded (Say for instance the connection was closed
before the stream management negotiation finished).

Future work would be to test different ways the
negotation could fail and to make sure that the
r and a stanzas are sent correctly.